fix epub publishing disabling resource build.
I can fix the publishResources flag to false because the hugo docs say:

"(...) Setting this to false will still publish Resources on demand (when a resource’s .Permalink or .RelPermalink is invoked from the templates) but will skip the others."

So whatever resource not used will not be published. The fix is to just reference a Permalink to images embedded as src tag.