Transform slug to title and move board image from Downloads folder.
I starting to use this script a lot and to improve it is a well paid time. So I made two more automation tasks: to use the slug as a base to the post title and if the board image in png format exists in the downloads folder to move it to the post folder. This way I don't have too much to do except to add a text before the image ('white move and win'). When there is more than one image, in other words, more than one moment to analyze, I need to add more than one shortcode to image and copy the second or third images from the downloads folder, but this is rarer than the single image scenario. So this change is a productivity advantage with little effort.