Edit.
Decided to update the post date, by some reasons:

Old posts updated deserve to be reposted in the news section. When changing a post it goes do draft again, and when removing from draft it goes back to news if I updated the date. To solve the history of the posts I must keep a track of what has changed in the newest version, making clear about the first date of the original post.

The draft can delay for many months until 'published'. The example is a book I am reading, stop reading, resume reading and one year has passed since the first draft. And now I am going to finish the post. I need to update because this is fair use of post history (my mind is from one year after I started reading, not one year before I finished the post).