Add new post about coffee.
I had an idea of not putting every post to the draft status. There are so small posts that I can review it in the first time and it is okay. This is going to give me a chance to write more during my copy & paste sections, where I just copy the text from the saved messages in Telegram and paste into my blog. I want to write more in the computer leveraging this time, and this instant review could be a way. Let's see.