Fix author image in author page.
For a couple hours I was tricked into thinking there was a bug in hugo to this
happens: the title from other posts was showing in a nonrelated post. This
behavior occured only in the parcial navbar, of course, the right title was in
the posts. However, there is two ways to render a partial into the layout:
partial ou partialCached. Apparently, partialCached does not parse partials for
every post.