Add section reposts.
This solves my problem with listing the news posts including those posts that
are old but that I keep up to date. I don't want to change the original post
date, but I would like that it showed up someway in some top listing.

In order to do this I created the repost page param that is the date of the last
repost. Based on that the new section lists the last reposts.

If for some reason I need to edit a old post as a draft I must remove the repost
tag.

Now thinking about it it solves the movies problem, also, since I can keep the
original watch date for a movie and keep up to date the last time I wrote about
it (because of some revision).