Update new publish date.
Now that I have a repost date does make more sense to fix the publish date as
the original post date, considering it is a new post. If a post was published
already so the repost date must be update, not the original date. That way the
post can show up in the repost section and not in the news section.