Create a post about segment tree.
In order to understand the Hacker Rank problem I had to watch a video and
study the code ChatGPT brought to me. In this way, finally I feel I can
write my own code to implement a binary tree and a segment tree, and I
know what this structure is useful.