Recap Hacker Rank array episode into Part 1.
This is because I am no more into doing so many interview tests right now.
This probably is going to be a archived project, but still that be defined
I just getting rid of this already reviewer for some times draft, following
my rule of thumb: do not let a draft too much time cooking.