Add support in awk.
This first migration from cpp to awk was to performance measure more than
a proof of concept. And I am not pleased with the result.

The cpp version took much less time than the awk version to render files
without write them (because the target is identical).

Perhaps to contruct and compile an awk interpreter for each post
is too much and I need to parse the entire blog at once.

If I could't solve the performance issue in awk I must consider to
use cpp entirely.