Transform blog posts to epub via awk.
I validate the epub thru epubcheck and despite invalid references to the
original template (easy to fix) there is a dozen os posts with invalid
characters (`&`, for instance).

The next step is get rid of those errors and keep all the epub
text files inside the awk script. After that the generation of epub
and mobi could be done via merged blog.txt.

Except that we don't have indexes, categories and tags yet =(