AWK script converts adoc to epub with 0 errors.
This version fixes some posts and the HTML conversion. Besides,
the new batch script uses the theme as a source to the origin files
that going to contribute to the final format.

The entire process, considering merge taking 1 seconds (or in the future
without the need to merge) takes about 15 seconds to generate the xhtml
posts and more about maximum 15 secs to pack the files to the zip that
forms the epub.

The kindle generation worked as well, but took much more time. The
size of the eboot is half the size of the original, though.