Fix references and errors and epub.
Big change here again. Since now there is a chapter for month nothing
better than use just the chapters as the main toc for the book. The
specific posts, tags and stuff will go all into the index when compiled.

This change let sumatra with an OK speed to open epub and mobi conversion
acceptable as well.

There are some improvements at sight:

 - order chronologically the chapters;
 - order chronologically the posts (could/must be done in merge?);
 - put at least the posts in index.