Start to use awk solution to ebook.
From this time the awk solution will be the main one, used to
publish an only ebook file to review drafts and server as the
entire blog in epub format, now the main one, and mobi format,
converted from epub using the legacy kindlegen.

The themes book and draft are now deprecated.