Remove old code.
Publish C++ code and awklib will not be used (awk now).
At the moment the conversion won't be to adoc, but simply a txt with
a title beginning with ´=´ to differenciate from markdown ´#´. The YAML
header still is converted to adoc format, though, because md does not
have a header.