Big migration from Hugo to Awk.
Now that the final merge was made I moved every post file and files related
to the hugo way to publish blogs and books to the archive folder. I refactored
the script files blog.cmd, book.cmd and clippings.py to pay attention to the
new paths.

Clippings are not in the source control anymore. Following the same logic
behind journal file, there is a clippings.txt file in home that is
populated incrementally and copied to Kindle every time a new book
version is available. This way we have at least two copies from clippings
and journal, "local" only.

And that's it. There is some minor changes to do, such as eliminate
categories and keep only tags, by example, but this kind of change does
not stop the writing process.