build a hardcoded mobi spine.
Epub conversion is not ideal because kindle mobi lack the flexibility of
all the indexes and links that were built in the hugo version. Besides, the
time to build a mobi from epub and from the source is the same. And finally,
I will use kindle for a very long time before it supports epub natively (
perhaps never).

So I am starting to build a script that similar to epub will construct a
mobi file directly from the txts I write. The first step is this commit,
that as was made in epub just outputs some hard coded model of what will
be the final mobi when capturing from the txt posts.

When the final script be ready the idea is to invert book generating
script from mobi as the default and epub as optional, just to publish
new versions to github or send to friends or test epubcheck.