starting to centralize content into Posts variable.
I discovered that it is not so straightforward to initialize an array
without creating an index. Using `a = {}` or `a = ()` does not work and
generate syntax error. Getting the `length()` of a new variable makes it
scalar.

So I am starting the Posts array with the index 0 as the blog metadata, not
the first post. So `Posts[0]["title"]` will contain the blog title, maybe
"author" would contain the author name and so on. When iterating the Posts
array to write the posts skip the index 0 and get the next index by calling
`newPostNumber = length(Posts)`, which will get automatically after BEGIN
the index 1.