get rid of minor local (global) variables in for loops.
The general idea is to use i, j, k, l, m, etc to name inner loops. The
advantage of this is to know that these variables can be ovewritten anytime
and that if there is a loop with a rare letter means there are too many
inner loops into this part.

And we do not let other variables to stay in the way of other functions and
be influenced by them.