first version with metadata.
This is a simple list that links slug with link. Yes, there is the bug
that identical slugs will be overriden. That's a shame.

The next step could be to fix this bug (don't know how with metadata)
or put other meta, such as titles. That could be tricky. Maybe removing all
the slugs could normalize slug generation and then its titles will match. Who
knows.