Detect unused explicit slugs
The purpose is to clear original text from slugs that were
created long time ago when this was important to them (e.g.
the movie reviews structure). If there is no need to obtain an
explicit slug to reference a post from others then there is no
need to keep this into the text.