Create initial function to extract original date from post
The idea is to use date field to contain future sections and also multiple
dates when the post were changed instead of the "v2" used now for miniviews.

It is not clear enough how globals are being used and need more refactory
into this part to understand the general idea and maybe do document a little.