Organize code to start doing entities
Now I understood the first thing to be refactored in the old code
is to discover the rules created in the AWK script and then convert
to entities in the new code. Only after that we can move forward to
a new parser and generator.

The AWK script is not only a Markdown parser. Actually, it is less
parser then reformatter to blog and ebook. We need to translate all
the logic behind tags organization, months posts and indexing to
the implied entities in the system.