Capture meta as paragraph lines next to heading
This is the first useful draft to really start to work. This
version pass only once in the file, parse all the document and
divide based on heading level 1. This heading is not included
in the document neither the first lines after the heading if they
are the next token. They are considered the metadata by the code
and insert it into a map.

I am still deciding what to do first but this is linked to the
decision of use pure C language or keep C++ interfaces. I am
positively tending to use C because cmark is already in C and
I don't see yet what can I possibly miss by not using C++.

Ref.: https://github.com/caloni/journal/issues/22