Generate months page with 90% equality to original
The main part is remaining is the tag navigation, where for this part
I was thinking in a performatic model where each entry process its
tags string as a pointer and add this pointer to the future sorted
navigation list with next and prev entry with same tag. It is inserted
by tag in a map maybe and when sort just change some elements just like
entries by date. This way the navigation part would be ready for each
entry because it is a reference to the same struct.

Something like that.