Write months page
This is the first template reconstructed from simple C language. Still
ideas of how to achieve a simple procedure to transform template pages
into final renderer. The main challenge is to make loops work based on
marks inside the templates.

In this version we have a hard coded snipped transformed in the exact
output from months page in the original awk script. A more flexible
solution should be insert a language inside the template. I do not
think this is feasible in a first draft. Let it this way for now.